+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.0-test2] - 2026-02-19
+### Added
+- **Inclusions / Exclusions:**
+  - Merged inclusion and Exclusion Manager
+  - moved Export feature to the Import / Export window (In BBMM Settings screen).
+    - Export exports both inclusions and exclusion lists. 
+    - Added "All" option to export one file will all inclusions / exclusions.
+
 ## [0.7.0-test1] - 2026-02-19
 ### Added
 - **Inclusions / Exclusions:**